Use of a counter instead of a input_number
https://www.home-assistant.io/integrations/counter/
A state of an input_number can be changed in the UI so that could possibly mess up the automations. 
Secondly, I also changed the layout of the automations a bit so it's more uniform and corresponding with the examples on the home-assistant.io website.
Lastly, also as an improvement I would suggest you could get also way with just two automations using the chooser. See 'make smarter automations, not more automations' from the Home Assistant Conference 2020 https://www.youtube.com/watch?v=OIkZWF5uGxk . I'm unsure if doing that would make it more or less comprehensive as an explanation, but the chooser is really powerful in HomeAssistant.