Wifi setup
I bought myself an ESP32 to see how the code works compared to an ESP8266. Didn't manage to get it into my network until I found that there is one wrong line in the Wifi setup of the code. After removing this line the ESP successfully connects to the Wifi.